rename timeout tracker and update docs for raw key pass-through
Rename the `timeout_event_generator` module to `stream_timeout_tracker` to better reflect its purpose as a standalone component.

Update the documentation and callback signatures to clarify that message keys are handled as raw pass-through objects (e.g., `bytes`) rather than being automatically decoded to strings. This ensures users are aware they must handle decoding within their custom callbacks if necessary.

Ref: SC-72221
